# Optimize reverse for 8 and 16-bit trivial types